+++ CHANGELOG.md
@@ -12,6 +12,12 @@ The REST API versions in the table are just for your information as the values a
 |<= 9.0.1|2
 
 <a name="11.7.0"></a>
+## 2019-03-27 Version [11.8.0](https://github.com/gooddata/gooddata-js/compare/v11.7.0...v11.8.0)
+
+- enhance AFM and execute-afm to support values instead of attribute element URIs in attribute filters used in
+  simple measures
+
+<a name="11.7.0"></a>
 ## 2019-03-25 Version [11.7.0](https://github.com/gooddata/gooddata-js/compare/v11.6.0...v11.7.0)
 
 - enhance AFM and execute-afm to support values instead of attribute element URIs in positive and negative attribute filters